fix: permission-checks

+++ server/controllers/favoriteLogo.controller.js
@@ -38,7 +38,7 @@ class FavoriteLogoController {
     }
   }
 
-  static async getByID(req, res) {
+  static async getById(req, res) {
     try {
       let favorite = await FavoriteLogoService.getById(req.params.id);
       res.json(favorite);
@@ -52,6 +52,12 @@ class FavoriteLogoController {
 
   static async create(req, res) {
     try {
+      if (req.user.id !== req.body.user_id) {
+        return res.status(403).json({
+          message: "You are not authorized to access this resource.",
+        });
+      }
+
       let favorite = await FavoriteLogoService.create(req.body);
       res.json(favorite);
     } catch (err) {
@@ -64,10 +70,14 @@ class FavoriteLogoController {
 
   static async update(req, res) {
     try {
-      let favorite = await FavoriteLogoService.updateById(
-        req.params.id,
-        req.body
-      );
+      let favorite = await FavoriteLogoService.getById(req.params.id);
+      if (req.user.id !== favorite.user_id) {
+        return res.status(403).json({
+          message: "You are not authorized to access this resource.",
+        });
+      }
+
+      favorite = await FavoriteLogoService.updateById(req.params.id, req.body);
       res.json(favorite);
     } catch (err) {
       logger.log("error", err.stack);
@@ -79,6 +89,13 @@ class FavoriteLogoController {
 
   static async delete(req, res) {
     try {
+      const favorite = await FavoriteLogoService.getById(req.params.id);
+      if (req.user.id !== favorite.user_id) {
+        return res.status(403).json({
+          message: "You are not authorized to access this resource.",
+        });
+      }
+
       await FavoriteLogoService.deleteById(req.params.id);
       res.json({});
     } catch (err) {
